@@ -7,12 +7,7 @@ import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.CoreMatchers.not;
 import static org.hamcrest.CoreMatchers.startsWith;
 import static org.hamcrest.Matchers.contains;
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertFalse;
-import static org.junit.Assert.assertNull;
-import static org.junit.Assert.assertSame;
-import static org.junit.Assert.assertThat;
-import static org.junit.Assert.assertTrue;
+import static org.junit.Assert.*;
 import static org.mockito.Matchers.matches;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.times;
@@ -37,9 +32,9 @@ import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
 
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import javax.servlet.http.HttpSession;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import jakarta.servlet.http.HttpSession;
 
 import org.junit.Rule;
 import org.junit.Test;
@@ -120,9 +115,9 @@ public class AuthTest {
 	 * @see com.onelogin.saml2.Auth
 	 */
 	@Test
-	public void testConstructor() throws IOException, SettingsException, Error {
+	public void testConstructor() throws Exception {
 		Auth auth = new Auth();
-		assertTrue(auth.getSettings() != null);
+        assertNotNull(auth.getSettings());
 
 		Saml2Settings settings = new SettingsBuilder().fromFile("onelogin.saml.properties").build();
 		assertEquals(settings.getIdpEntityId(), auth.getSettings().getIdpEntityId());
@@ -141,9 +136,9 @@ public class AuthTest {
 	 * @see com.onelogin.saml2.Auth
 	 */
 	@Test
-	public void testConstructorWithFilename() throws IOException, SettingsException, Error {
+	public void testConstructorWithFilename() throws Exception {
 		Auth auth = new Auth("config/config.min.properties");
-		assertTrue(auth.getSettings() != null);
+        assertNotNull(auth.getSettings());
 
 		Saml2Settings settings = new SettingsBuilder().fromFile("config/config.min.properties").build();
 		assertEquals(settings.getIdpEntityId(), auth.getSettings().getIdpEntityId());
@@ -165,12 +160,12 @@ public class AuthTest {
      * @see com.onelogin.saml2.Auth
      */
 	@Test
-	public void testConstructorWithFilenameAndKeyStore() throws IOException, SettingsException, Error, NoSuchAlgorithmException, CertificateException, KeyStoreException, UnrecoverableKeyException {
+	public void testConstructorWithFilenameAndKeyStore() throws Exception {
         
 		Auth auth = new Auth("config/config.min.properties", getKeyStoreSettings());
-		assertTrue(auth.getSettings() != null);
-		assertTrue(auth.getSettings().getSPcert() != null);
-		assertTrue(auth.getSettings().getSPkey() != null);
+        assertNotNull(auth.getSettings());
+        assertNotNull(auth.getSettings().getSPcert());
+        assertNotNull(auth.getSettings().getSPkey());
 		
 		Saml2Settings settings = new SettingsBuilder().fromFile("config/config.min.properties", getKeyStoreSettings()).build();
 		assertEquals(settings.getSPcert(), auth.getSettings().getSPcert());
@@ -189,14 +184,14 @@ public class AuthTest {
 	 * @see com.onelogin.saml2.Auth
 	 */
 	@Test
-	public void testConstructorWithReqRes() throws IOException, SettingsException, URISyntaxException, Error {
+	public void testConstructorWithReqRes() throws Exception {
 		HttpServletRequest request = mock(HttpServletRequest.class);
 		HttpServletResponse response = mock(HttpServletResponse.class);
 		String samlResponseEncoded = Util.getFileAsString("data/responses/response1.xml.base64");
 		when(request.getParameterMap()).thenReturn(singletonMap("SAMLResponse", new String[]{samlResponseEncoded}));
 
 		Auth auth = new Auth(request, response);
-		assertTrue(auth.getSettings() != null);
+        assertNotNull(auth.getSettings());
 
 		Saml2Settings settings = new SettingsBuilder().fromFile("onelogin.saml.properties").build();
 		assertEquals(settings.getIdpEntityId(), auth.getSettings().getIdpEntityId());
@@ -218,14 +213,14 @@ public class AuthTest {
 	 * @see com.onelogin.saml2.Auth
 	 */
 	@Test
-	public void testConstructorWithReqResAndKeyStore() throws IOException, SettingsException, URISyntaxException, Error, KeyStoreException, NoSuchAlgorithmException, CertificateException {
+	public void testConstructorWithReqResAndKeyStore() throws Exception, KeyStoreException, NoSuchAlgorithmException, CertificateException {
 		HttpServletRequest request = mock(HttpServletRequest.class);
 		HttpServletResponse response = mock(HttpServletResponse.class);
 
 		Auth auth = new Auth(getKeyStoreSettings(), request, response);
-		assertTrue(auth.getSettings() != null);
-		assertTrue(auth.getSettings().getSPcert() != null);
-		assertTrue(auth.getSettings().getSPkey() != null);
+        assertNotNull(auth.getSettings());
+        assertNotNull(auth.getSettings().getSPcert());
+        assertNotNull(auth.getSettings().getSPkey());
 
 		Saml2Settings settings = new SettingsBuilder().fromFile("onelogin.saml.properties", getKeyStoreSettings()).build();
 		assertEquals(settings.getSPkey(), auth.getSettings().getSPkey());
@@ -244,14 +239,14 @@ public class AuthTest {
 	 * @see com.onelogin.saml2.Auth
 	 */
 	@Test
-	public void testConstructorWithFilenameReqRes() throws IOException, SettingsException, URISyntaxException, Error {
+	public void testConstructorWithFilenameReqRes() throws Exception {
 		HttpServletRequest request = mock(HttpServletRequest.class);
 		HttpServletResponse response = mock(HttpServletResponse.class);
 		String samlResponseEncoded = Util.getFileAsString("data/responses/response1.xml.base64");
 		when(request.getParameterMap()).thenReturn(singletonMap("SAMLResponse", new String[]{samlResponseEncoded}));
 
 		Auth auth = new Auth("config/config.min.properties", request, response);
-		assertTrue(auth.getSettings() != null);
+        assertNotNull(auth.getSettings());
 
 		Saml2Settings settings = new SettingsBuilder().fromFile("config/config.min.properties").build();
 		assertEquals(settings.getIdpEntityId(), auth.getSettings().getIdpEntityId());
@@ -270,7 +265,7 @@ public class AuthTest {
 	 * @see com.onelogin.saml2.Auth
 	 */
 	@Test
-	public void testConstructorWithSettingsReqRes() throws IOException, SettingsException, URISyntaxException, Error {
+	public void testConstructorWithSettingsReqRes() throws Exception {
 		HttpServletRequest request = mock(HttpServletRequest.class);
 		HttpServletResponse response = mock(HttpServletResponse.class);
 		String samlResponseEncoded = Util.getFileAsString("data/responses/response1.xml.base64");
@@ -294,7 +289,7 @@ public class AuthTest {
 	 * @see com.onelogin.saml2.Auth
 	 */
 	@Test
-	public void testConstructorInvalidSettings() throws IOException, SettingsException, URISyntaxException, Error {
+	public void testConstructorInvalidSettings() throws Exception {
 		HttpServletRequest request = mock(HttpServletRequest.class);
 		HttpServletResponse response = mock(HttpServletResponse.class);
 		String samlResponseEncoded = Util.getFileAsString("data/responses/response1.xml.base64");
@@ -317,7 +312,7 @@ public class AuthTest {
 	 * @see com.onelogin.saml2.Auth#getSettings
 	 */
 	@Test
-	public void testGetSettings() throws IOException, SettingsException, Error {
+	public void testGetSettings() throws Exception {
 		Saml2Settings settings = new SettingsBuilder().fromFile("onelogin.saml.properties").build();
 		Auth auth = new Auth();
 		assertEquals(settings.getIdpEntityId(), auth.getSettings().getIdpEntityId());
@@ -345,7 +340,7 @@ public class AuthTest {
 	 * @see com.onelogin.saml2.Auth#setStrict
 	 */
 	@Test
-	public void testSetStrict() throws IOException, SettingsException, URISyntaxException, Error {
+	public void testSetStrict() throws Exception {
 		Auth auth = new Auth();
 
 		auth.setStrict(false);
@@ -366,7 +361,7 @@ public class AuthTest {
 	 * @see com.onelogin.saml2.Auth#isDebugActive
 	 */
 	@Test
-	public void testIsDebugActive() throws IOException, SettingsException, URISyntaxException, Error {
+	public void testIsDebugActive() throws Exception {
 		HttpServletResponse response = mock(HttpServletResponse.class);
 		HttpServletRequest request = mock(HttpServletRequest.class);
 		String samlResponseEncoded = Util.getFileAsString("data/responses/response1.xml.base64");
@@ -394,7 +389,7 @@ public class AuthTest {
 	 * @see com.onelogin.saml2.Auth#getSSOurl
 	 */
 	@Test
-	public void testGetSSOurl() throws URISyntaxException, IOException, SettingsException, Error {
+	public void testGetSSOurl() throws Exception {
 		HttpServletRequest request = mock(HttpServletRequest.class);
 		HttpServletResponse response = mock(HttpServletResponse.class);
 		String samlResponseEncoded = Util.getFileAsString("data/responses/response1.xml.base64");
@@ -417,7 +412,7 @@ public class AuthTest {
 	 * @see com.onelogin.saml2.Auth#getSLOurl
 	 */
 	@Test
-	public void testGetSLOurl() throws URISyntaxException, IOException, SettingsException, Error {
+	public void testGetSLOurl() throws Exception {
 		HttpServletRequest request = mock(HttpServletRequest.class);
 		HttpServletResponse response = mock(HttpServletResponse.class);
 		String samlResponseEncoded = Util.getFileAsString("data/responses/response1.xml.base64");
@@ -441,7 +436,7 @@ public class AuthTest {
 	 * @see com.onelogin.saml2.Auth#getSLOResponseUrl
 	 */
 	@Test
-	public void testGetSLOResponseUrl() throws URISyntaxException, IOException, SettingsException, Error {
+	public void testGetSLOResponseUrl() throws Exception {
 		HttpServletRequest request = mock(HttpServletRequest.class);
 		HttpServletResponse response = mock(HttpServletResponse.class);
 		String samlResponseEncoded = Util.getFileAsString("data/responses/response1.xml.base64");
@@ -464,7 +459,7 @@ public class AuthTest {
 	 * @see com.onelogin.saml2.Auth#getSLOResponseUrl
 	 */
 	@Test
-	public void testGetSLOResponseUrlNull() throws URISyntaxException, IOException, SettingsException, Error {
+	public void testGetSLOResponseUrlNull() throws Exception {
 		HttpServletRequest request = mock(HttpServletRequest.class);
 		HttpServletResponse response = mock(HttpServletResponse.class);
 		String samlResponseEncoded = Util.getFileAsString("data/responses/response1.xml.base64");
@@ -1290,7 +1285,7 @@ public class AuthTest {
 	 * @see com.onelogin.saml2.Auth#login
 	 */
 	@Test
-	public void testLogin() throws IOException, SettingsException, URISyntaxException, Error {
+	public void testLogin() throws Exception {
 		HttpServletRequest request = mock(HttpServletRequest.class);
 		HttpServletResponse response = mock(HttpServletResponse.class);
 		when(request.getScheme()).thenReturn("http");
@@ -1318,7 +1313,7 @@ public class AuthTest {
 	 * @see com.onelogin.saml2.Auth#login
 	 */
 	@Test
-	public void testLoginWithRelayState() throws IOException, SettingsException, URISyntaxException, Error {
+	public void testLoginWithRelayState() throws Exception {
 		HttpServletRequest request = mock(HttpServletRequest.class);
 		HttpServletResponse response = mock(HttpServletResponse.class);
 		when(request.getScheme()).thenReturn("http");
@@ -1347,7 +1342,7 @@ public class AuthTest {
 	 * @see com.onelogin.saml2.Auth#login
 	 */
 	@Test
-	public void testLoginWithoutRelayState() throws IOException, SettingsException, URISyntaxException, Error {
+	public void testLoginWithoutRelayState() throws Exception {
 		HttpServletRequest request = mock(HttpServletRequest.class);
 		HttpServletResponse response = mock(HttpServletResponse.class);
 		when(request.getScheme()).thenReturn("http");
@@ -1378,7 +1373,7 @@ public class AuthTest {
 	 * @see com.onelogin.saml2.Auth#login
 	 */
 	@Test
-	public void testLoginWithExtraParameters() throws IOException, SettingsException, URISyntaxException, Error {
+	public void testLoginWithExtraParameters() throws Exception {
 		HttpServletRequest request = mock(HttpServletRequest.class);
 		HttpServletResponse response = mock(HttpServletResponse.class);
 		when(request.getScheme()).thenReturn("http");
@@ -1409,7 +1404,7 @@ public class AuthTest {
 	 * @see com.onelogin.saml2.Auth#login
 	 */
 	@Test
-	public void testLoginStay() throws IOException, SettingsException, URISyntaxException, Error {
+	public void testLoginStay() throws Exception {
 		HttpServletRequest request = mock(HttpServletRequest.class);
 		HttpServletResponse response = mock(HttpServletResponse.class);
 		when(request.getScheme()).thenReturn("http");
@@ -1443,7 +1438,7 @@ public class AuthTest {
 	 * @see com.onelogin.saml2.Auth#login
 	 */
 	@Test
-	public void testLoginSubject() throws IOException, SettingsException, URISyntaxException, Error {
+	public void testLoginSubject() throws Exception {
 		HttpServletRequest request = mock(HttpServletRequest.class);
 		HttpServletResponse response = mock(HttpServletResponse.class);
 		when(request.getScheme()).thenReturn("http");
@@ -1492,7 +1487,7 @@ public class AuthTest {
 	 * @see com.onelogin.saml2.Auth#login
 	 */
 	@Test
-	public void testLoginSignedFail() throws IOException, SettingsException, URISyntaxException, Error {
+	public void testLoginSignedFail() throws Exception {
 		HttpServletRequest request = mock(HttpServletRequest.class);
 		HttpServletResponse response = mock(HttpServletResponse.class);
 		when(request.getScheme()).thenReturn("http");
@@ -1521,7 +1516,7 @@ public class AuthTest {
 	 * @see com.onelogin.saml2.Auth#login
 	 */
 	@Test
-	public void testLoginSigned() throws IOException, SettingsException, URISyntaxException, Error {
+	public void testLoginSigned() throws Exception {
 		HttpServletRequest request = mock(HttpServletRequest.class);
 		HttpServletResponse response = mock(HttpServletResponse.class);
 		when(request.getScheme()).thenReturn("http");
@@ -1555,7 +1550,7 @@ public class AuthTest {
 	 * @see com.onelogin.saml2.Auth#logout
 	 */
 	@Test
-	public void testLogout() throws IOException, SettingsException, XMLEntityException, Error {
+	public void testLogout() throws Exception {
 		HttpServletRequest request = mock(HttpServletRequest.class);
 		HttpServletResponse response = mock(HttpServletResponse.class);
 		when(request.getScheme()).thenReturn("http");
@@ -1584,7 +1579,7 @@ public class AuthTest {
 	 * @see com.onelogin.saml2.Auth#logout
 	 */
 	@Test
-	public void testLogoutWithExtraParameters() throws IOException, SettingsException, XMLEntityException, Error {
+	public void testLogoutWithExtraParameters() throws Exception {
 		HttpServletRequest request = mock(HttpServletRequest.class);
 		HttpServletResponse response = mock(HttpServletResponse.class);
 		when(request.getScheme()).thenReturn("http");
@@ -1614,7 +1609,7 @@ public class AuthTest {
 	 * @see com.onelogin.saml2.Auth#logout
 	 */
 	@Test
-	public void testLogoutWithRelayState() throws IOException, SettingsException, XMLEntityException, Error {
+	public void testLogoutWithRelayState() throws Exception {
 		HttpServletRequest request = mock(HttpServletRequest.class);
 		HttpServletResponse response = mock(HttpServletResponse.class);
 		when(request.getScheme()).thenReturn("http");
@@ -1644,7 +1639,7 @@ public class AuthTest {
 	 * @see com.onelogin.saml2.Auth#logout
 	 */
 	@Test
-	public void testLogoutWithoutRelayState() throws IOException, SettingsException, XMLEntityException, Error {
+	public void testLogoutWithoutRelayState() throws Exception {
 		HttpServletRequest request = mock(HttpServletRequest.class);
 		HttpServletResponse response = mock(HttpServletResponse.class);
 		when(request.getScheme()).thenReturn("http");
@@ -1676,7 +1671,7 @@ public class AuthTest {
 	 * @see com.onelogin.saml2.Auth#logout
 	 */
 	@Test
-	public void testLogoutStay() throws IOException, SettingsException, XMLEntityException, Error {
+	public void testLogoutStay() throws Exception {
 		HttpServletRequest request = mock(HttpServletRequest.class);
 		HttpServletResponse response = mock(HttpServletResponse.class);
 		when(request.getScheme()).thenReturn("http");
@@ -1710,7 +1705,7 @@ public class AuthTest {
 	 * @see com.onelogin.saml2.Auth#logout
 	 */
 	@Test
-	public void testLogoutSignedFail() throws IOException, SettingsException, XMLEntityException, Error {
+	public void testLogoutSignedFail() throws Exception {
 		HttpServletRequest request = mock(HttpServletRequest.class);
 		HttpServletResponse response = mock(HttpServletResponse.class);
 		when(request.getScheme()).thenReturn("http");
@@ -1739,7 +1734,7 @@ public class AuthTest {
 	 * @see com.onelogin.saml2.Auth#logout
 	 */
 	@Test
-	public void testLogoutSigned() throws IOException, SettingsException, XMLEntityException, Error {
+	public void testLogoutSigned() throws Exception {
 		HttpServletRequest request = mock(HttpServletRequest.class);
 		HttpServletResponse response = mock(HttpServletResponse.class);
 		when(request.getScheme()).thenReturn("http");
@@ -1773,7 +1768,7 @@ public class AuthTest {
 	 * @see com.onelogin.saml2.Auth#buildRequestSignature
 	 */
 	@Test
-	public void testBuildRequestSignatureInvalidSP() throws URISyntaxException, IOException, SettingsException, Error {
+	public void testBuildRequestSignatureInvalidSP() throws Exception {
 		String deflatedEncodedAuthNRequest = Util.getFileAsString("data/requests/authn_request.xml.deflated.base64");
 		String relayState = "http://example.com";
 		String signAlgorithm = "http://www.w3.org/2000/09/xmldsig#rsa-sha1";
@@ -1797,7 +1792,7 @@ public class AuthTest {
 	 * @see com.onelogin.saml2.Auth#buildRequestSignature
 	 */
 	@Test
-	public void testBuildRequestSignatureRsaSha1() throws URISyntaxException, IOException, SettingsException, Error {
+	public void testBuildRequestSignatureRsaSha1() throws Exception {
 		String deflatedEncodedAuthNRequest = Util.getFileAsString("data/requests/authn_request.xml.deflated.base64");
 		String relayState = "http://example.com";
 		String signAlgorithm = Constants.RSA_SHA1;
@@ -1823,7 +1818,7 @@ public class AuthTest {
 	 * @see com.onelogin.saml2.Auth#buildRequestSignature
 	 */
 	@Test(expected=IllegalArgumentException.class)
-	public void testBuildRequestSignatureDsaSha1() throws URISyntaxException, IOException, SettingsException, Error {
+	public void testBuildRequestSignatureDsaSha1() throws Exception {
 		String deflatedEncodedAuthNRequest = Util.getFileAsString("data/requests/authn_request.xml.deflated.base64");
 		String relayState = "http://example.com";
 		String signAlgorithm = Constants.DSA_SHA1;
@@ -1844,7 +1839,7 @@ public class AuthTest {
 	 * @see com.onelogin.saml2.Auth#buildRequestSignature
 	 */
 	@Test
-	public void testBuildRequestSignatureRsaSha256() throws URISyntaxException, IOException, SettingsException, Error {
+	public void testBuildRequestSignatureRsaSha256() throws Exception {
 		String deflatedEncodedAuthNRequest = Util.getFileAsString("data/requests/authn_request.xml.deflated.base64");
 		String relayState = "http://example.com";
 		String signAlgorithm = Constants.RSA_SHA256;
@@ -1867,7 +1862,7 @@ public class AuthTest {
 	 * @see com.onelogin.saml2.Auth#buildRequestSignature
 	 */
 	@Test
-	public void testBuildRequestSignatureRsaSha384() throws URISyntaxException, IOException, SettingsException, Error {
+	public void testBuildRequestSignatureRsaSha384() throws Exception {
 		String deflatedEncodedAuthNRequest = Util.getFileAsString("data/requests/authn_request.xml.deflated.base64");
 		String relayState = "http://example.com";
 		String signAlgorithm = Constants.RSA_SHA384;
@@ -1890,7 +1885,7 @@ public class AuthTest {
 	 * @see com.onelogin.saml2.Auth#buildRequestSignature
 	 */
 	@Test
-	public void testBuildRequestSignatureRsaSha512() throws URISyntaxException, IOException, SettingsException, Error {
+	public void testBuildRequestSignatureRsaSha512() throws Exception {
 		String deflatedEncodedAuthNRequest = Util.getFileAsString("data/requests/authn_request.xml.deflated.base64");
 		String relayState = "http://example.com";		
 		String signAlgorithm = Constants.RSA_SHA512;
@@ -1913,7 +1908,7 @@ public class AuthTest {
 	 * @see com.onelogin.saml2.Auth#buildResponseSignature
 	 */
 	@Test
-	public void testBuildResponseSignatureRsaSha1() throws URISyntaxException, IOException, SettingsException, Error {
+	public void testBuildResponseSignatureRsaSha1() throws Exception {
 		String deflatedEncodedLogoutResponse = Util.getFileAsString("data/logout_responses/logout_response_deflated.xml.base64");
 		String relayState = "http://example.com";
 		String signAlgorithm = Constants.RSA_SHA1;
@@ -1939,7 +1934,7 @@ public class AuthTest {
 	 * @see com.onelogin.saml2.Auth#buildResponseSignature
 	 */
 	@Test(expected=IllegalArgumentException.class)
-	public void testBuildResponseSignatureDsaSha1() throws URISyntaxException, IOException, SettingsException, Error {
+	public void testBuildResponseSignatureDsaSha1() throws Exception {
 		String deflatedEncodedLogoutResponse = Util.getFileAsString("data/logout_responses/logout_response_deflated.xml.base64");
 		String relayState = "http://example.com";
 		String signAlgorithm = Constants.DSA_SHA1;
@@ -1960,7 +1955,7 @@ public class AuthTest {
 	 * @see com.onelogin.saml2.Auth#buildResponseSignature
 	 */
 	@Test
-	public void testBuildResponseSignatureRsaSha256() throws URISyntaxException, IOException, SettingsException, Error {
+	public void testBuildResponseSignatureRsaSha256() throws Exception {
 		String deflatedEncodedLogoutResponse = Util.getFileAsString("data/logout_responses/logout_response_deflated.xml.base64");
 		String relayState = "http://example.com";
 		String signAlgorithm = Constants.RSA_SHA256;
@@ -1983,7 +1978,7 @@ public class AuthTest {
 	 * @see com.onelogin.saml2.Auth#buildResponseSignature
 	 */
 	@Test
-	public void testBuildResponseSignatureRsaSha384() throws URISyntaxException, IOException, SettingsException, Error {
+	public void testBuildResponseSignatureRsaSha384() throws Exception {
 		String deflatedEncodedLogoutResponse = Util.getFileAsString("data/logout_responses/logout_response_deflated.xml.base64");
 		String relayState = "http://example.com";
 		String signAlgorithm = Constants.RSA_SHA384;
@@ -2006,7 +2001,7 @@ public class AuthTest {
 	 * @see com.onelogin.saml2.Auth#buildResponseSignature
 	 */
 	@Test
-	public void testBuildResponseSignatureRsaSha512() throws URISyntaxException, IOException, SettingsException, Error {
+	public void testBuildResponseSignatureRsaSha512() throws Exception {
 		String deflatedEncodedLogoutResponse = Util.getFileAsString("data/logout_responses/logout_response_deflated.xml.base64");
 		String relayState = "http://example.com";		
 		String signAlgorithm = Constants.RSA_SHA512;
@@ -2028,7 +2023,7 @@ public class AuthTest {
 	 * @see com.onelogin.saml2.Auth#buildSignature
 	 */
 	@Test
-	public void testBuildSignature() throws URISyntaxException, IOException, SettingsException, Error {
+	public void testBuildSignature() throws Exception {
 		String deflatedEncodedAuthNRequest = Util.getFileAsString("data/requests/authn_request.xml.deflated.base64");
 		String deflatedEncodedLogoutResponse = Util.getFileAsString("data/logout_responses/logout_response_deflated.xml.base64");
 		String relayState = "http://example.com";		
@@ -2091,7 +2086,7 @@ public class AuthTest {
 	 * @see com.onelogin.saml2.Auth#getLastRequestXML
 	 */
 	@Test
-	public void testGetLastAuthNRequest() throws IOException, SettingsException, Error {
+	public void testGetLastAuthNRequest() throws Exception {
 		HttpServletRequest request = mock(HttpServletRequest.class);
 		HttpServletResponse response = mock(HttpServletResponse.class);
 		Saml2Settings settings = new SettingsBuilder().fromFile("config/config.min.properties").build();
@@ -2117,7 +2112,7 @@ public class AuthTest {
 	 * @see com.onelogin.saml2.Auth#getLastRequestXML
 	 */
 	@Test
-	public void testGetLastLogoutRequestSent() throws IOException, SettingsException, XMLEntityException, Error {
+	public void testGetLastLogoutRequestSent() throws Exception {
 		HttpServletRequest request = mock(HttpServletRequest.class);
 		HttpServletResponse response = mock(HttpServletResponse.class);
 		Saml2Settings settings = new SettingsBuilder().fromFile("config/config.min.properties").build();
