@@ -10,6 +10,16 @@ import com.onelogin.saml2.model.RequestedAttribute;
  * A class that stores AttributeConsumingService
  */
 public class AttributeConsumingService {
+
+	/**
+	 * Service Index
+	 */
+	private final int index;
+	/**
+	 * Whether this service is the default one
+	 */
+	private Boolean isDefault;
+
 	/**
      * Service Name
      */
@@ -21,6 +31,11 @@ public class AttributeConsumingService {
 	private final String serviceDescription;
 
 	/**
+	 * Language used for service name and description
+	 */
+	private String lang;
+
+	/**
      * Requested Attributes
      */
 	private final List<RequestedAttribute> requestedAttributes;
@@ -28,15 +43,58 @@ public class AttributeConsumingService {
 	/**
 	 * Constructor
 	 *
+	 * @param index
+	 *              int. Service index
+	 * @param isDefault
+	 *              boolean. Whether it's the default attribute consuming service
+	 *
 	 * @param serviceName
 	 *              String. Service Name
 	 * @param serviceDescription
 	 *              String. Service Description
 	 */
-	public AttributeConsumingService(String serviceName, String serviceDescription) {
+	public AttributeConsumingService(int index, Boolean isDefault, String serviceName, String serviceDescription, String lang) {
+		this.index = index;
+		this.isDefault = isDefault;
 		this.serviceName = serviceName != null? serviceName : "";
-		this.serviceDescription = serviceDescription != null? serviceDescription : "";
-		this.requestedAttributes = new ArrayList<RequestedAttribute>();
+		this.serviceDescription = serviceDescription;
+		this.lang = lang != null? lang: "en";
+		this.requestedAttributes = new ArrayList<>();
+	}
+
+
+	/**
+	 * Constructor. Service name and description are assumed to be in English.
+	 *
+	 * @param index
+	 *              int. Service index
+	 * @param isDefault
+	 *              boolean. Whether it's the default attribute consuming service
+	 * @param serviceName
+	 *              String. Service Name
+	 * @param serviceDescription
+	 *              String. Service Description
+	 */
+	public AttributeConsumingService(int index, Boolean isDefault, String serviceName, String serviceDescription) {
+		this(index, isDefault, serviceName, serviceDescription, null);
+	}
+
+	/**
+	 * Constructor for a non-default attribute consuming service with index
+	 * <code>1</code> and service name and descriptions in English.
+	 * <p>
+	 * Mainly kept for backward compatibility, this constructor can be used when an
+	 * only attribute consuming service is required. Please also note that, to
+	 * maintain full backward compatibility, if the service description is
+	 * <code>null</code> this constructor will set is as an empty string.
+	 *
+	 * @param serviceName
+	 *              String. Service Name
+	 * @param serviceDescription
+	 *              String. Service Description; if <code>null</code>, an empty string will be set
+	 */
+	public AttributeConsumingService(String serviceName, String serviceDescription) {
+		this(1, null, serviceName, serviceDescription != null? serviceDescription : "", null);
 	}
 
 	/**
@@ -68,4 +126,32 @@ public class AttributeConsumingService {
 		return requestedAttributes;
 	}
 
+	/**
+	 * @return int the service index
+	 */
+	public final int getIndex() {
+		return index;
+	}
+
+	/**
+	 * @return boolean whether this is the default attribute consuming service
+	 */
+	public final Boolean isDefault() {
+		return isDefault;
+	}
+
+	/**
+	 * @return string the language in which service name and description are written
+	 */
+	public String getLang() {
+		return lang;
+	}
+
+	public void setLang(String lang) {
+		this.lang = lang;
+	}
+
+	public void setDefault(Boolean aDefault) {
+		isDefault = aDefault;
+	}
 }
