@@ -112,7 +112,7 @@ public class MetadataTest {
 	 * @see com.onelogin.saml2.settings.Metadata
 	 */
 	@Test
-	public void testMetadataExpiration() throws IOException, CertificateEncodingException, Error {
+	public void testMetadataExpiration() throws Exception {
 		Saml2Settings settings = new SettingsBuilder().fromFile("config/config.min.properties").build();
 
 		Metadata metadataObj = new Metadata(settings, null, null);
@@ -142,7 +142,7 @@ public class MetadataTest {
 	 * @see com.onelogin.saml2.settings.Metadata#toContactsXml
 	 */
 	@Test
-	public void testToContactsXml() throws IOException, CertificateEncodingException, Error {
+	public void testToContactsXml() throws Exception {
 		Saml2Settings settings = getSettingFromAllProperties();
 		Metadata metadataObj = new Metadata(settings);
 		String metadataStr = metadataObj.getMetadataString();
@@ -171,7 +171,7 @@ public class MetadataTest {
 	 * @see com.onelogin.saml2.settings.Metadata#toContactsXml
 	 */
 	@Test
-	public void testToContactsXmlSpecialChars() throws IOException, CertificateEncodingException, Error {
+	public void testToContactsXmlSpecialChars() throws Exception {
 		Saml2Settings settings = getSettingFromAllSpecialCharsProperties();
 		Metadata metadataObj = new Metadata(settings);
 		String metadataStr = metadataObj.getMetadataString();
@@ -198,7 +198,7 @@ public class MetadataTest {
 	 * @see com.onelogin.saml2.settings.Metadata#toContactsXml
 	 */
 	@Test
-	public void testToContactsXmlLegacy() throws IOException, CertificateEncodingException, Error {
+	public void testToContactsXmlLegacy() throws Exception {
 		Saml2Settings settings = getSettingFromAllProperties();
 		Metadata metadataObj = new Metadata(settings);
 		String metadataStr = metadataObj.getMetadataString();
@@ -227,7 +227,7 @@ public class MetadataTest {
 	 * @see com.onelogin.saml2.settings.Metadata#toContactsXml
 	 */
 	@Test
-	public void testToContactsXmlLegacySpecialChars() throws IOException, CertificateEncodingException, Error {
+	public void testToContactsXmlLegacySpecialChars() throws Exception {
 		Saml2Settings settings = getSettingFromAllSpecialCharsProperties();
 		Metadata metadataObj = new Metadata(settings);
 		String metadataStr = metadataObj.getMetadataString();
@@ -247,7 +247,7 @@ public class MetadataTest {
 	 * @see com.onelogin.saml2.settings.Metadata#toOrganizationXml
 	 */
 	@Test
-	public void testToOrganizationXml() throws IOException, CertificateEncodingException, Error {
+	public void testToOrganizationXml() throws Exception {
 		Saml2Settings settings = getSettingFromAllProperties();
 		Metadata metadataObj = new Metadata(settings);
 		String metadataStr = metadataObj.getMetadataString();
@@ -271,7 +271,7 @@ public class MetadataTest {
 	 * @see com.onelogin.saml2.settings.Metadata#toOrganizationXml
 	 */
 	@Test
-	public void testToNonLocalizedOrganizationXml() throws IOException, CertificateEncodingException, Error {
+	public void testToNonLocalizedOrganizationXml() throws Exception {
 		Saml2Settings settings = new SettingsBuilder().fromFile("config/config.org.properties").build();
 		Metadata metadataObj = new Metadata(settings);
 		String metadataStr = metadataObj.getMetadataString();
@@ -296,7 +296,7 @@ public class MetadataTest {
 	 * @see com.onelogin.saml2.settings.Metadata#toOrganizationXml
 	 */
 	@Test
-	public void testToLocalizedOrganizationXml() throws IOException, CertificateEncodingException, Error {
+	public void testToLocalizedOrganizationXml() throws Exception {
 		Saml2Settings settings = new SettingsBuilder().fromFile("config/config.org.localized.properties").build();
 		Metadata metadataObj = new Metadata(settings);
 		String metadataStr = metadataObj.getMetadataString();
@@ -322,7 +322,7 @@ public class MetadataTest {
 	 * @see com.onelogin.saml2.settings.Metadata#toOrganizationXml
 	 */
 	@Test
-	public void testToOrganizationXmlSpecialChars() throws IOException, CertificateEncodingException, Error {
+	public void testToOrganizationXmlSpecialChars() throws Exception {
 		Saml2Settings settings = getSettingFromAllSpecialCharsProperties();
 		Metadata metadataObj = new Metadata(settings);
 		String metadataStr = metadataObj.getMetadataString();
@@ -340,7 +340,7 @@ public class MetadataTest {
 	 * @see com.onelogin.saml2.settings.Metadata#toSLSXml
 	 */
 	@Test
-	public void testToSLSXml() throws IOException, CertificateEncodingException, Error {
+	public void testToSLSXml() throws Exception {
 		Saml2Settings settings = getSettingFromAllProperties();
 		Metadata metadataObj = new Metadata(settings);
 		String metadataStr = metadataObj.getMetadataString();
@@ -367,7 +367,7 @@ public class MetadataTest {
 	 * @see com.onelogin.saml2.settings.Metadata#toSLSXml
 	 */
 	@Test
-	public void testToSLSXmlSpecialChars() throws IOException, CertificateEncodingException, Error {
+	public void testToSLSXmlSpecialChars() throws Exception {
 		Saml2Settings settings = getSettingFromAllSpecialCharsProperties();
 		Metadata metadataObj = new Metadata(settings);
 		String metadataStr = metadataObj.getMetadataString();
@@ -386,7 +386,7 @@ public class MetadataTest {
 	 * @see com.onelogin.saml2.settings.Metadata#toX509KeyDescriptorsXML
 	 */
 	@Test
-	public void testToX509KeyDescriptorsXML() throws IOException, CertificateEncodingException, Error {
+	public void testToX509KeyDescriptorsXML() throws Exception {
 		Saml2Settings settings = getSettingFromAllProperties();
 		Metadata metadataObj = new Metadata(settings);
 		String metadataStr = metadataObj.getMetadataString();
@@ -421,7 +421,7 @@ public class MetadataTest {
 	 * @see com.onelogin.saml2.settings.Metadata#toX509KeyDescriptorsXML
 	 */
 	@Test
-	public void testToX509KeyDescriptorsXMLEncryption() throws IOException, CertificateEncodingException, Error {
+	public void testToX509KeyDescriptorsXMLEncryption() throws Exception {
 		Saml2Settings settings = getSettingFromAllProperties();
 		String keyDescriptorEncStr = "<md:KeyDescriptor use=\"encryption\"><ds:KeyInfo xmlns:ds=\"http://www.w3.org/2000/09/xmldsig#\"><ds:X509Data><ds:X509Certificate>MIICeDCCAeGgAwIBAgIBADANBgkqhkiG9w0BAQ0FADBZMQswCQYDVQQGEwJ1czET";
 
@@ -458,8 +458,92 @@ public class MetadataTest {
 	 * @throws Error
 	 * @see com.onelogin.saml2.settings.Metadata#getAttributeConsumingServiceXml
 	 */
+
+	public void testToAttributeConsumingServicesXmlLegacy() throws Exception {
+		Saml2Settings settings = getSettingFromAllProperties();
+
+		AttributeConsumingService attributeConsumingService = new AttributeConsumingService(0, true, "Test Service", "Test Service Desc", "en");
+		RequestedAttribute requestedAttribute = new RequestedAttribute("Email", "Email", true, "urn:oasis:names:tc:SAML:2.0:attrname-format:uri", null);
+		RequestedAttribute requestedAttribute2 = new RequestedAttribute("FirstName", null, true, "urn:oasis:names:tc:SAML:2.0:attrname-format:uri", null);
+		RequestedAttribute requestedAttribute3 = new RequestedAttribute("LastName", null, true, "urn:oasis:names:tc:SAML:2.0:attrname-format:uri", null);
+
+		attributeConsumingService.addRequestedAttribute(requestedAttribute);
+		attributeConsumingService.addRequestedAttribute(requestedAttribute2);
+		attributeConsumingService.addRequestedAttribute(requestedAttribute3);
+
+		Metadata metadataObj = new Metadata(settings, null, null, attributeConsumingService);
+		String metadataStr = metadataObj.getMetadataString();
+
+		String headerStr = "<md:AttributeConsumingService index=\"0\" isDefault=\"true\">";
+		String sNameStr = "<md:ServiceName xml:lang=\"en\">Test Service</md:ServiceName>";
+		String sDescStr = "<md:ServiceDescription xml:lang=\"en\">Test Service Desc</md:ServiceDescription>";
+		String reqAttr1Str = "<md:RequestedAttribute Name=\"Email\" NameFormat=\"urn:oasis:names:tc:SAML:2.0:attrname-format:uri\" FriendlyName=\"Email\" isRequired=\"true\" />";
+		String reqAttr2Str = "<md:RequestedAttribute Name=\"FirstName\" NameFormat=\"urn:oasis:names:tc:SAML:2.0:attrname-format:uri\" isRequired=\"true\" />";
+		String reqAttr3Str = "<md:RequestedAttribute Name=\"LastName\" NameFormat=\"urn:oasis:names:tc:SAML:2.0:attrname-format:uri\" isRequired=\"true\" />";
+		String footerStr = "</md:AttributeConsumingService>";
+
+		assertThat(metadataStr, containsString(headerStr));
+		assertThat(metadataStr, containsString(sNameStr));
+		assertThat(metadataStr, containsString(sDescStr));
+		assertThat(metadataStr, containsString(reqAttr1Str));
+		assertThat(metadataStr, containsString(reqAttr2Str));
+		assertThat(metadataStr, containsString(reqAttr3Str));
+		assertThat(metadataStr, containsString(footerStr));
+	}
+	/**
+	 * Tests the toAttributeConsumingServicesXml method of Metadata when an Attribute Consuming Service
+	 * is specified at construction time
+	 * <p>
+	 * Case: Attribute Consuming Service definition contains special chars.
+	 *
+	 * @throws IOException
+	 * @throws CertificateEncodingException
+	 * @throws Error
+	 * @see com.onelogin.saml2.settings.Metadata#toAttributeConsumingServicesXml
+	 */
 	@Test
-	public void testGetAttributeConsumingServiceXml() throws IOException, CertificateEncodingException, Error {
+	public void testToAttributeConsumingServicesXmlLegacySpecialChars() throws Exception {
+		Saml2Settings settings = getSettingFromAllProperties();
+
+		AttributeConsumingService attributeConsumingService = new AttributeConsumingService(0, true, "T&st Service", "T&st Service Desc");
+		RequestedAttribute requestedAttribute = new RequestedAttribute("Email", "Email \"address\"", true, "urn:oasis:names:tc:SAML:2.0:attrname-format:uri", null);
+		RequestedAttribute requestedAttribute2 = new RequestedAttribute("FirstName&LastName", null, true, "urn:oasis:names:tc:SAML:2.0:attrname-format:uri", null);
+
+		attributeConsumingService.addRequestedAttribute(requestedAttribute);
+		attributeConsumingService.addRequestedAttribute(requestedAttribute2);
+
+		Metadata metadataObj = new Metadata(settings, null, null, attributeConsumingService);
+		String metadataStr = metadataObj.getMetadataString();
+
+		String headerStr = "<md:AttributeConsumingService index=\"0\" isDefault=\"true\">";
+		String sNameStr = "<md:ServiceName xml:lang=\"en\">T&amp;st Service</md:ServiceName>";
+		String sDescStr = "<md:ServiceDescription xml:lang=\"en\">T&amp;st Service Desc</md:ServiceDescription>";
+		String reqAttr1Str = "<md:RequestedAttribute Name=\"Email\" NameFormat=\"urn:oasis:names:tc:SAML:2.0:attrname-format:uri\" FriendlyName=\"Email &quot;address&quot;\" isRequired=\"true\" />";
+		String reqAttr2Str = "<md:RequestedAttribute Name=\"FirstName&amp;LastName\" NameFormat=\"urn:oasis:names:tc:SAML:2.0:attrname-format:uri\" isRequired=\"true\" />";
+		String footerStr = "</md:AttributeConsumingService>";
+
+		assertThat(metadataStr, containsString(headerStr));
+		assertThat(metadataStr, containsString(sNameStr));
+		assertThat(metadataStr, containsString(sDescStr));
+		assertThat(metadataStr, containsString(reqAttr1Str));
+		assertThat(metadataStr, containsString(reqAttr2Str));
+		assertThat(metadataStr, containsString(footerStr));
+	}
+
+	/**
+	 * Tests the toAttributeConsumingServicesXml method of Metadata when an Attribute Consuming Service
+	 * is specified at construction time
+	 * <p>
+	 * Case: Single non-default AttributeConsumingService with no explicit index.
+	 *
+	 * @throws IOException
+	 * @throws CertificateEncodingException
+	 * @throws Error
+	 * @see com.onelogin.saml2.settings.Metadata#toAttributeConsumingServicesXml
+	 */
+
+	@Test
+	public void testToAttributeConsumingServiceXmlNoIndexLegacy() throws Exception {
 		Saml2Settings settings = getSettingFromAllProperties();
 
 		AttributeConsumingService attributeConsumingService = new AttributeConsumingService("Test Service", "Test Service Desc");
@@ -502,7 +586,7 @@ public class MetadataTest {
 	 * @see com.onelogin.saml2.settings.Metadata#getAttributeConsumingServiceXml
 	 */
 	@Test
-	public void testGetAttributeConsumingServiceXmlSpecialChars() throws IOException, CertificateEncodingException, Error {
+	public void testToAttributeConsumingServiceXmlNoIndexLegacySpecialChars() throws Exception {
 		Saml2Settings settings = getSettingFromAllProperties();
 
 		AttributeConsumingService attributeConsumingService = new AttributeConsumingService("T&st Service", "T&st Service Desc");
@@ -541,10 +625,10 @@ public class MetadataTest {
 	 * @see com.onelogin.saml2.settings.Metadata#getAttributeConsumingServiceXml
 	 */
 	@Test
-	public void testGetAttributeConsumingServiceXmlWithMultipleAttributeValue() throws IOException, CertificateEncodingException, Error {
+	public void testToAttributeConsumingServiceXmlWithMultipleAttributeValueLegacy() throws Exception {
 		Saml2Settings settings = getSettingFromAllProperties();
 
-		AttributeConsumingService attributeConsumingService = new AttributeConsumingService("Test Service", "Test Service Desc");
+		AttributeConsumingService attributeConsumingService = new AttributeConsumingService(0, true, "Test Service", "Test Service Desc", "en");
 		List<String> attrValues = new ArrayList<String>();
 		attrValues.add("userType");
 		attrValues.add("admin");
@@ -557,7 +641,7 @@ public class MetadataTest {
 		Metadata metadataObj = new Metadata(settings, null, null, attributeConsumingService);
 		String metadataStr = metadataObj.getMetadataString();
 
-		String headerStr = "<md:AttributeConsumingService index=\"1\">";
+		String headerStr = "<md:AttributeConsumingService index=\"0\" isDefault=\"true\">";
 		String sNameStr = "<md:ServiceName xml:lang=\"en\">Test Service</md:ServiceName>";
 		String sDescStr = "<md:ServiceDescription xml:lang=\"en\">Test Service Desc</md:ServiceDescription>";
 		String reqAttr1Str = "<md:RequestedAttribute Name=\"userType\" NameFormat=\"urn:oasis:names:tc:SAML:2.0:attrname-format:basic\" isRequired=\"false\">";
@@ -587,7 +671,7 @@ public class MetadataTest {
 	 * @see com.onelogin.saml2.settings.Metadata#getAttributeConsumingServiceXml
 	 */
 	@Test
-	public void testGetAttributeConsumingServiceXmlWithMultipleAttributeValueSpecialChars() throws IOException, CertificateEncodingException, Error {
+	public void testToAttributeConsumingServiceXmlWithMultipleAttributeValueLegacySpecialChars() throws Exception {
 		Saml2Settings settings = getSettingFromAllProperties();
 
 		AttributeConsumingService attributeConsumingService = new AttributeConsumingService("T&st Service", "T&st Service Desc");
@@ -620,6 +704,143 @@ public class MetadataTest {
 		assertThat(metadataStr, containsString(reqAttr1Attr2Str));
 		assertThat(metadataStr, containsString(reqAttr2Str));
 		assertThat(metadataStr, containsString(footerStr));
+	}
+
+	/**
+	 * Tests the toAttributeConsumingServicesXml method of Metadata
+	 * <p>
+	 * Case: single Attribute Consuming Service specified in settings.
+	 *
+	 * @throws IOException
+	 * @throws CertificateEncodingException
+	 * @throws Error
+	 * @see com.onelogin.saml2.settings.Metadata#toAttributeConsumingServicesXml
+	 */
+	@Test
+	public void testToAttributeConsumingServiceXmlSingleService() throws Exception {
+		Saml2Settings settings = getSettingFromAllProperties();
+
+		Metadata metadataObj = new Metadata(settings, null, null);
+		String metadataStr = metadataObj.getMetadataString();
+
+		String headerStr = "<md:AttributeConsumingService index=\"1\">";
+		String sNameStr = "<md:ServiceName xml:lang=\"en\">My service</md:ServiceName>";
+		String sDescStr = "<md:ServiceDescription xml:lang=\"en\">My service description</md:ServiceDescription>";
+		String reqAttr1Str = "<md:RequestedAttribute Name=\"Email\" NameFormat=\"urn:oasis:names:tc:SAML:1.1:nameid-format:emailAddress\" FriendlyName=\"E-mail address\" isRequired=\"true\">";
+		String reqAttr1Atr1Str = "<saml:AttributeValue xmlns:saml=\"urn:oasis:names:tc:SAML:2.0:assertion\">foo@example.org</saml:AttributeValue>";
+		String reqAttr1Attr2Str = "<saml:AttributeValue xmlns:saml=\"urn:oasis:names:tc:SAML:2.0:assertion\">bar@example.org</saml:AttributeValue>";
+		String reqAttr2Str = "<md:RequestedAttribute Name=\"FirstName\" />";
+		String footerStr = "</md:AttributeConsumingService>";
+
+		assertThat(metadataStr, containsString(headerStr));
+		assertThat(metadataStr, containsString(sNameStr));
+		assertThat(metadataStr, containsString(sDescStr));
+		assertThat(metadataStr, containsString(reqAttr1Str));
+		assertThat(metadataStr, containsString(reqAttr1Atr1Str));
+		assertThat(metadataStr, containsString(reqAttr1Attr2Str));
+		assertThat(metadataStr, containsString(reqAttr2Str));
+		assertThat(metadataStr, containsString(footerStr));
+	}
+
+	/**
+	 * Tests the toAttributeConsumingServicesXml method of Metadata
+	 * <p>
+	 * Case: single Attribute Consuming Service specified in settings.
+	 *
+	 * @throws IOException
+	 * @throws CertificateEncodingException
+	 * @throws Error
+	 * @see com.onelogin.saml2.settings.Metadata#toAttributeConsumingServicesXml
+	 */
+	@Test
+	public void testToAttributeConsumingServiceXmlSingleACSSpecialChars() throws Exception {
+		Saml2Settings settings = getSettingFromAllSpecialCharsProperties();
+
+		Metadata metadataObj = new Metadata(settings, null, null);
+		String metadataStr = metadataObj.getMetadataString();
+
+		String headerStr = "<md:AttributeConsumingService index=\"1\">";
+		String sNameStr = "<md:ServiceName xml:lang=\"&amp;n\">My s&amp;rvice</md:ServiceName>";
+		String sDescStr = "<md:ServiceDescription xml:lang=\"&amp;n\">My s&amp;rvice description</md:ServiceDescription>";
+		String reqAttr1Str = "<md:RequestedAttribute Name=\"Email\" NameFormat=\"urn:oasis:names:tc:SAML:1.1:nameid-format:emailAddress\" FriendlyName=\"E-mail &quot;address&quot;\" isRequired=\"true\">";
+		String reqAttr1Atr1Str = "<saml:AttributeValue xmlns:saml=\"urn:oasis:names:tc:SAML:2.0:assertion\">foo&amp;bar@example.org</saml:AttributeValue>";
+		String reqAttr1Attr2Str = "<saml:AttributeValue xmlns:saml=\"urn:oasis:names:tc:SAML:2.0:assertion\">bar@example.org</saml:AttributeValue>";
+		String reqAttr2Str = "<md:RequestedAttribute Name=\"FirstName&amp;LastName\" />";
+		String footerStr = "</md:AttributeConsumingService>";
+
+		assertThat(metadataStr, containsString(headerStr));
+		assertThat(metadataStr, containsString(sNameStr));
+		assertThat(metadataStr, containsString(sDescStr));
+		assertThat(metadataStr, containsString(reqAttr1Str));
+		assertThat(metadataStr, containsString(reqAttr1Atr1Str));
+		assertThat(metadataStr, containsString(reqAttr1Attr2Str));
+		assertThat(metadataStr, containsString(reqAttr2Str));
+		assertThat(metadataStr, containsString(footerStr));
+	}
+
+	/**
+	 * Tests the toAttributeConsumingServicesXml method of Metadata
+	 * Case: single Attribute Consuming Service specified in settings
+	 *
+	 * @throws IOException
+	 * @throws CertificateEncodingException
+	 * @throws Error
+	 * @see com.onelogin.saml2.settings.Metadata#toAttributeConsumingServicesXml
+	 */
+	@Test
+	public void testToAttributeConsumingServiceXmlMultiServices() throws Exception {
+		Saml2Settings settings = getSettingFromAllPropertiesMultiAttributeConsumingServices();
+
+		Metadata metadataObj = new Metadata(settings, null, null);
+		String metadataStr = metadataObj.getMetadataString();
+
+		String header1Str = "<md:AttributeConsumingService index=\"0\">";
+		String sName1Str = "<md:ServiceName xml:lang=\"en\">Just e-mail</md:ServiceName>";
+		String reqAttr11Str = "<md:RequestedAttribute Name=\"Email\" NameFormat=\"urn:oasis:names:tc:SAML:1.1:nameid-format:emailAddress\" FriendlyName=\"E-mail address\" isRequired=\"true\">";
+		String reqAttr11Atr1Str = "<saml:AttributeValue xmlns:saml=\"urn:oasis:names:tc:SAML:2.0:assertion\">foo@example.org</saml:AttributeValue>";
+		String reqAttr11Attr2Str = "<saml:AttributeValue xmlns:saml=\"urn:oasis:names:tc:SAML:2.0:assertion\">bar@example.org</saml:AttributeValue>";
+		String footer1Str = "</md:AttributeConsumingService>";
+
+		String header2Str = "<md:AttributeConsumingService index=\"1\" isDefault=\"true\">";
+		String sName2Str = "<md:ServiceName xml:lang=\"it\">Anagrafica &quot;completa&quot;</md:ServiceName>";
+		String sDesc2Str = "<md:ServiceDescription xml:lang=\"it\">Dati anagrafici &quot;completi&quot;</md:ServiceDescription>";
+		String reqAttr21Str = "<md:RequestedAttribute Name=\"Birth date\" />";
+		String reqAttr22Str = "<md:RequestedAttribute Name=\"First &amp; Last Name\" isRequired=\"true\" />";
+		String footer2Str = "</md:AttributeConsumingService>";
+
+		assertThat(metadataStr, containsString(header1Str));
+		assertThat(metadataStr, containsString(sName1Str));
+		assertThat(metadataStr, containsString(reqAttr11Str));
+		assertThat(metadataStr, containsString(reqAttr11Atr1Str));
+		assertThat(metadataStr, containsString(reqAttr11Attr2Str));
+		assertThat(metadataStr, containsString(footer1Str));
+
+		assertThat(metadataStr, containsString(header2Str));
+		assertThat(metadataStr, containsString(sName2Str));
+		assertThat(metadataStr, containsString(sDesc2Str));
+		assertThat(metadataStr, containsString(reqAttr21Str));
+		assertThat(metadataStr, containsString(reqAttr22Str));
+		assertThat(metadataStr, containsString(footer2Str));
+
+		// properties for the single non-indexed ACS must NOT be present in this case
+		String sNameStr = "<md:ServiceName xml:lang=\"en\">My service</md:ServiceName>";
+		String sDescStr = "<md:ServiceDescription xml:lang=\"en\">My service description</md:ServiceDescription>";
+		String reqAttr1Str = "<md:RequestedAttribute Name=\"Email_Wrong\" NameFormat=\"urn:oasis:names:tc:SAML:1.1:nameid-format:emailAddress\" FriendlyName=\"E-mail address\" isRequired=\"true\">";
+		String reqAttr1Atr1Str = "<saml:AttributeValue xmlns:saml=\"urn:oasis:names:tc:SAML:2.0:assertion\">foo_wrong@example.org</saml:AttributeValue>";
+		String reqAttr1Attr2Str = "<saml:AttributeValue xmlns:saml=\"urn:oasis:names:tc:SAML:2.0:assertion\">bar_wrong@example.org</saml:AttributeValue>";
+		String reqAttr2Str = "<md:RequestedAttribute Name=\"FirstName_Wrong\" />";
+
+		assertThat(metadataStr, not(containsString(sNameStr)));
+		assertThat(metadataStr, not(containsString(sDescStr)));
+		assertThat(metadataStr, not(containsString(reqAttr1Str)));
+		assertThat(metadataStr, not(containsString(reqAttr1Atr1Str)));
+		assertThat(metadataStr, not(containsString(reqAttr1Attr2Str)));
+		assertThat(metadataStr, not(containsString(reqAttr2Str)));
+	}
+
+
+	private Saml2Settings getSettingFromAllPropertiesMultiAttributeConsumingServices() throws Exception {
+		return new SettingsBuilder().fromFile("config/config.all_multi_attribute_consuming_services.properties").build();
 	}
 
 	/**
@@ -671,7 +892,7 @@ public class MetadataTest {
 	 * @see com.onelogin.saml2.settings.Metadata#signMetadata
 	 */
 	@Test
-	public void testSigngeneratedMetadata() throws Error, IOException, GeneralSecurityException, XPathExpressionException, XMLSecurityException {
+	public void testSigngeneratedMetadata() throws Exception {
 		Saml2Settings settings = getSettingFromAllProperties();
 		String certString = Util.getFileAsString("data/customPath/certs/sp.crt");
 		X509Certificate cert = Util.loadCert(certString);
@@ -699,16 +920,16 @@ public class MetadataTest {
 		assertEquals(digestAlgorithmSha512, digest_method_metadata_signed.getAttributes().getNamedItem("Algorithm").getNodeValue());
 	}
 
-	private Saml2Settings getSettingFromAllProperties() throws Error, IOException {
+	private Saml2Settings getSettingFromAllProperties() throws Exception {
 		return new SettingsBuilder().fromFile("config/config.all.properties").build();
 	}
 
-	private Saml2Settings getSettingFromAllSpecialCharsProperties() throws Error, IOException {
+	private Saml2Settings getSettingFromAllSpecialCharsProperties() throws Exception {
 		return new SettingsBuilder().fromFile("config/config.all_specialchars.properties").build();
 	}
 
 	@Test
-	public void shouldIncludeValidUntilAndDuration() throws CertificateEncodingException, Error, IOException {
+	public void shouldIncludeValidUntilAndDuration() throws Exception {
 		//given
 		Saml2Settings saml2Settings = getSettingFromAllProperties();
 
@@ -728,7 +949,7 @@ public class MetadataTest {
 	}
 
 	@Test
-	public void shouldIgnoreValidUntil() throws CertificateEncodingException, Error, IOException {
+	public void shouldIgnoreValidUntil() throws Exception {
 		//given
 		Saml2Settings saml2Settings = getSettingFromAllProperties();
 
@@ -747,7 +968,7 @@ public class MetadataTest {
 	}
 
 	@Test
-	public void shouldIgnoreCacheDuration() throws CertificateEncodingException, Error, IOException {
+	public void shouldIgnoreCacheDuration() throws Exception {
 		//given
 		Saml2Settings saml2Settings = getSettingFromAllProperties();
 
